test(canon): cover dependency resolution for a symlinked node_modules store
